rejigger metrics file layout
