+++ myapp/config/initializers/sidekiq.rb
@@ -49,6 +49,9 @@ Sidekiq::CurrentAttributes.persist(Myapp::Current) # Your AS::CurrentAttributes 
 
 # Sidekiq.transactional_push!
 
-Sidekiq.configure_server do |config|
-  Sidekiq::Metrics.track(config)
-end
+# create a label based on the shorthash and subject line of the latest commit in git.
+# WARNING: you only want to run this ONCE! If this runs on boot for 20 different Sidekiq processes,
+# you will get 20 different deploy marks in Redis! Instead this should go into the script
+# that runs your deploy, e.g. your capistrano script.
+label = `git log -1 --format="%h %s"`.strip
+Sidekiq::Metrics::Deploy.new.mark(label: label)
